Move unit test into epoch_rewards_hasher sub-submodule
